+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.7.1] - 2020-03-16
 
 ### Added
 
@@ -150,6 +150,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.5.0]: https://github.com/dogmatiq/configkit/releases/v0.5.0
 [0.6.0]: https://github.com/dogmatiq/configkit/releases/v0.6.0
 [0.7.0]: https://github.com/dogmatiq/configkit/releases/v0.7.0
+[0.7.1]: https://github.com/dogmatiq/configkit/releases/v0.7.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD